Restructure application.
