+++ src/components/App/App.jsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { combineReducers, createStore, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import { Provider } from 'react-redux';
-import { reducer as simulation } from './state';
-import Simulator from './Simulator';
+import { reducer as simulation } from '../../lib/state';
+import Simulator from '../Simulator/Simulator';
 import './App.css';
 
 const rootReducer = combineReducers({
@@ -14,8 +14,10 @@ const middlewares = applyMiddleware(thunk);
 
 const store = createStore(rootReducer, middlewares);
 
-export default () => (
+const App = () => (
   <Provider store={store}>
     <Simulator />
   </Provider>
 );
+
+export default App;
