+++ src/components/Rule/Rule.css
