+++ src/components/Rule/Rule.jsx
