+++ src/components/Rules/Rules.jsx
@@ -1,9 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import Rule, { RulePropType } from './Rule';
-import RuleForm from './RuleForm';
-import { addRule, removeRule } from './state';
+import Rule, { RulePropType } from '../Rule/Rule';
+import RuleForm from '../Rule/RuleForm';
 
 const Rules = ({ rules, addRule, removeRule }) => (
   <div className="rules">
@@ -24,11 +22,4 @@ Rules.propTypes = {
   removeRule: PropTypes.func.isRequired,
 };
 
-const mapStateToProps = state => ({
-  rules: state.simulation.rules,
-});
-
-export default connect(
-  mapStateToProps,
-  { removeRule, addRule }
-)(Rules);
+export default Rules;
