+++ src/components/Rules/index.js
@@ -1,0 +1,12 @@
+import { connect } from 'react-redux';
+import { addRule, removeRule } from '../../lib/state';
+import Rules from './Rules';
+
+const mapStateToProps = state => ({
+  rules: state.simulation.rules,
+});
+
+export default connect(
+  mapStateToProps,
+  { removeRule, addRule }
+)(Rules);
